+++ discovery-provider/plugins/pedalboard/apps/relay/src/routes/health.ts
@@ -1,6 +1,10 @@
 import { NextFunction, Request, Response } from "express";
 
-export const healthCheck = async (req: Request, res: Response, next: NextFunction) => {
+export const healthCheck = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
   res.send({ status: "up" });
-  next()
+  next();
 };
